+++ wetterdienst/ui/restapi.py
@@ -16,7 +16,13 @@ from wetterdienst.core.scalar.request import ScalarRequestCore
 from wetterdienst.exceptions import ProviderError
 from wetterdienst.metadata.columns import Columns
 from wetterdienst.ui.cli import get_api
-from wetterdienst.ui.core import get_stations, get_values, set_logging_level
+from wetterdienst.ui.core import (
+    get_interpolate,
+    get_stations,
+    get_summarize,
+    get_values,
+    set_logging_level,
+)
 from wetterdienst.util.cli import read_list, setup_logging
 
 app = FastAPI(debug=False)
@@ -58,6 +64,8 @@ def index():
                 <li><a href=restapi/coverage>coverage</a></li>
                 <li><a href=restapi/stations>stations</a></li>
                 <li><a href=restapi/values>values</a></li>
+                <li><a href=restapi/interpolate>interpolation</a></li>
+                <li><a href=restapi/summarize>summary</a></li>
             </ul>
             <h4>Producer</h4>
             {PRODUCER_NAME} - <a href="{PRODUCER_LINK}">{PRODUCER_LINK}</a></li>
@@ -65,6 +73,8 @@ def index():
             <ul>
             <li><a href="restapi/stations?provider=dwd&network=observation&parameter=kl&resolution=daily&period=recent&all=true">DWD Observation stations</a></li>
             <li><a href="restapi/values?provider=dwd&network=observation&parameter=kl&resolution=daily&period=recent&station=00011">DWD Observation values</a></li>
+            <li><a href="restapi/interpolate?provider=dwd&network=observation&parameter=kl&resolution=daily&station=00071&date=1986-10-31/1986-11-01">DWD Observation values</a></li>
+            <li><a href="restapi/summarize?provider=dwd&network=observation&parameter=kl&resolution=daily&station=00071&date=1986-10-31/1986-11-01">DWD Observation values</a></li>
             </ul>
         </body>
     </html>
@@ -257,6 +267,9 @@ def values(
     :param humanize:
     :param tidy:        Whether to return data in tidy format. Default: True.
     :param si_units:
+    :param skip_empty:
+    :param skip_threshold:
+    :param dropna:
     :param pretty:
     :param debug:
     :return:
@@ -321,6 +334,232 @@ def values(
             skip_threshold=skip_threshold,
             dropna=dropna,
             tidy=tidy,
+            humanize=humanize,
+        )
+    except Exception as e:
+        log.exception(e)
+
+        return HTTPException(status_code=404, detail=str(e))
+
+    indent = None
+    if pretty:
+        indent = 4
+
+    output = values_.df
+
+    output[Columns.DATE.value] = output[Columns.DATE.value].apply(lambda ts: ts.isoformat())
+
+    output = output.replace({np.NaN: None, pd.NA: None})
+
+    output = output.to_dict(orient="records")
+
+    output = make_json_response(output, api.provider)
+
+    output = json.dumps(output, indent=indent, ensure_ascii=False)
+
+    return Response(content=output, media_type="application/json")
+
+
+@app.get("/restapi/interpolate")
+def interpolate(
+    provider: str = Query(default=None),
+    network: str = Query(default=None),
+    parameter: str = Query(default=None),
+    resolution: str = Query(default=None),
+    period: str = Query(default=None),
+    use_nearby_station_until_km: float = Query(default=1.0),
+    date: str = Query(default=None),
+    issue: str = Query(default="latest"),
+    station: str = Query(default=None),
+    coordinates: str = Query(default=None),
+    sql_values: str = Query(alias="sql-values", default=None),
+    humanize: bool = Query(default=True),
+    si_units: bool = Query(alias="si-units", default=True),
+    pretty: bool = Query(default=False),
+    debug: bool = Query(default=False),
+):
+    """
+    Wrapper around get_interpolate to provide results via restapi
+
+    :param provider:
+    :param network:        string for network of provider
+    :param parameter:   Observation measure
+    :param resolution:  Frequency/granularity of measurement interval
+    :param period:      Recent or historical files
+    :param date:        Date or date range
+    :param issue:
+    :param station:
+    :param coordinates:
+    :param sql_values:
+    :param humanize:
+    :param si_units:
+    :param pretty:
+    :param debug:
+    :return:
+    """
+    # TODO: Add geojson support
+    fmt = "json"
+
+    if provider is None or network is None:
+        raise HTTPException(
+            status_code=400,
+            detail="Query arguments 'provider' and 'network' are required",
+        )
+
+    if parameter is None or resolution is None:
+        raise HTTPException(
+            status_code=400,
+            detail="Query arguments 'parameter', 'resolution' " "and 'date' are required",
+        )
+
+    if fmt not in ("json", "geojson"):
+        raise HTTPException(
+            status_code=400,
+            detail="format argument must be one of json, geojson",
+        )
+
+    set_logging_level(debug)
+
+    try:
+        api: ScalarRequestCore = Wetterdienst(provider, network)
+    except ProviderError:
+        return HTTPException(
+            status_code=404,
+            detail=f"Given combination of provider and network not available. "
+            f"Choose provider and network from {Wetterdienst.discover()}",
+        )
+
+    parameter = read_list(parameter)
+    if period:
+        period = read_list(period)
+    if station:
+        station = read_list(station)
+
+    try:
+        values_ = get_interpolate(
+            api=api,
+            parameter=parameter,
+            resolution=resolution,
+            date=date,
+            issue=issue,
+            period=period,
+            station_id=station,
+            coordinates=coordinates,
+            sql_values=sql_values,
+            si_units=si_units,
+            humanize=humanize,
+            use_nearby_station_until_km=use_nearby_station_until_km,
+        )
+    except Exception as e:
+        log.exception(e)
+
+        return HTTPException(status_code=404, detail=str(e))
+
+    indent = None
+    if pretty:
+        indent = 4
+
+    output = values_.df
+
+    output[Columns.DATE.value] = output[Columns.DATE.value].apply(lambda ts: ts.isoformat())
+
+    output = output.replace({np.NaN: None, pd.NA: None})
+
+    output = output.to_dict(orient="records")
+
+    output = make_json_response(output, api.provider)
+
+    output = json.dumps(output, indent=indent, ensure_ascii=False)
+
+    return Response(content=output, media_type="application/json")
+
+
+@app.get("/restapi/summarize")
+def summarize(
+    provider: str = Query(default=None),
+    network: str = Query(default=None),
+    parameter: str = Query(default=None),
+    resolution: str = Query(default=None),
+    period: str = Query(default=None),
+    date: str = Query(default=None),
+    issue: str = Query(default="latest"),
+    station: str = Query(default=None),
+    coordinates: str = Query(default=None),
+    sql_values: str = Query(alias="sql-values", default=None),
+    humanize: bool = Query(default=True),
+    si_units: bool = Query(alias="si-units", default=True),
+    pretty: bool = Query(default=False),
+    debug: bool = Query(default=False),
+):
+    """
+    Wrapper around get_summarize to provide results via restapi
+
+    :param provider:
+    :param network:        string for network of provider
+    :param parameter:   Observation measure
+    :param resolution:  Frequency/granularity of measurement interval
+    :param period:      Recent or historical files
+    :param date:        Date or date range
+    :param issue:
+    :param station:
+    :param coordinates:
+    :param sql_values:
+    :param humanize:
+    :param si_units:
+    :param pretty:
+    :param debug:
+    :return:
+    """
+    # TODO: Add geojson support
+    fmt = "json"
+
+    if provider is None or network is None:
+        raise HTTPException(
+            status_code=400,
+            detail="Query arguments 'provider' and 'network' are required",
+        )
+
+    if parameter is None or resolution is None:
+        raise HTTPException(
+            status_code=400,
+            detail="Query arguments 'parameter', 'resolution' " "and 'date' are required",
+        )
+
+    if fmt not in ("json", "geojson"):
+        raise HTTPException(
+            status_code=400,
+            detail="format argument must be one of json, geojson",
+        )
+
+    set_logging_level(debug)
+
+    try:
+        api: ScalarRequestCore = Wetterdienst(provider, network)
+    except ProviderError:
+        return HTTPException(
+            status_code=404,
+            detail=f"Given combination of provider and network not available. "
+            f"Choose provider and network from {Wetterdienst.discover()}",
+        )
+
+    parameter = read_list(parameter)
+    if period:
+        period = read_list(period)
+    if station:
+        station = read_list(station)
+
+    try:
+        values_ = get_summarize(
+            api=api,
+            parameter=parameter,
+            resolution=resolution,
+            date=date,
+            issue=issue,
+            period=period,
+            station_id=station,
+            coordinates=coordinates,
+            sql_values=sql_values,
+            si_units=si_units,
             humanize=humanize,
         )
     except Exception as e: